Create README.md
Just started looking into new career paths and I'd previously taken classes in programming and coding before, but only introductory and never got to experience it in a way that made sense or enjoyable. I'm hoping that by taking this at my own pace and focusing on applications that I would enjoy, I might be able to find a renewed interest and even develop it into a proper skill set that I not only enjoy but can also make a living doing.